+++ frontend/Idle/src/components/common/content/Car3D.jsx
@@ -1,19 +1,6 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { styled } from "styled-components";
 import palette from "styles/palette";
-
-function preloadImage(src = null) {
-  return new Promise((resolve, reject) => {
-    const img = new Image()
-    img.onload = function () {
-      resolve(img)
-    }
-    img.onerror = img.onabort = function () {
-      reject(src)
-    }
-    img.src = src
-  })
-}
 
 function Car3D({ data = null }) {
   const [currentImg, setCurrentImage] = useState(0);
@@ -21,9 +8,6 @@ function Car3D({ data = null }) {
   const [beforeX, setBeforeX] = useState(0);
   let imgCount = data === null ? 0 : data[0]?.carImgUrls.length - 1;
 
-  useEffect(() => {
-    data === null ? "" : data[0]?.carImgUrls?.map((item) => { preloadImage(item.imgUrl) })
-  }, [data])
   function turnLeft() {
     if (currentImg === imgCount) {
       setCurrentImage(0);
@@ -40,11 +24,10 @@ function Car3D({ data = null }) {
   }
   function turnCar(e) {
     if (isMouseDown && e.clientX != beforeX) {
-      console.log(e.clientX - beforeX);
-      if (e.clientX - beforeX > 1) {
+      if (e.clientX - beforeX > 2) {
         turnRight()
         setBeforeX(e.clientX);
-      } else if (beforeX - e.clientX > 1) {
+      } else if (beforeX - e.clientX > 2) {
         turnLeft()
         setBeforeX(e.clientX);
       }
